product-image-search/requirements.txt
fastapi[standard]
openai
uvicorn
python-dotenv
python-multipart
aiohttp
pillow
httpx
torch
qdrant_client
transformers
torchvision